Refactored auth functions into blueprint
- Submodule handles all login/logout requests.
- Renamed auth.py to auth_utils.py, moved endpoint-specific functions to
  helpers.py and left everything that might be used in many places in
  auth_utils.py.
