+++ RuddockWebsite/modules/auth/__init__.py
@@ -1,0 +1,4 @@
+from flask import Blueprint
+blueprint = Blueprint('auth', __name__, template_folder='templates')
+
+import RuddockWebsite.modules.auth.routes
