+++ RuddockWebsite/modules/auth/templates/forgot_password.html
@@ -1,7 +1,7 @@
 {% extends "layout.html" %}
 {% block body %}
   <h2>Reset my password!</h2>
-  <form action="{{ url_for('forgot_password_submit') }}" method="post">
+  <form action="{{ url_for('auth.forgot_password_submit') }}" method="post">
     <table class="form">
       <tr>
         <td>Username:</td>
