+++ RuddockWebsite/modules/auth/templates/reset_password.html
@@ -1,7 +1,7 @@
 {% extends "layout.html" %}
 {% block body %}
   <h2>Reset my password!</h2>
-  <form action="{{ url_for('reset_password_submit', reset_key=reset_key) }}" method="post">
+  <form action="{{ url_for('auth.reset_password_submit', reset_key=reset_key) }}" method="post">
     <table class="form">
       <tr>
         <td>Username:</td>
