--- RuddockWebsite/templates/password_change.html
@@ -1,26 +1,0 @@
-{% extends "layout.html" %}
-{% block body %}
-  <h2>Change my password!</h2>
-  <form action="" method="post">
-    <table class="form">
-      <tr>
-        <td>Username</td>
-        <td>{{session['username']}}</td>
-      </tr>
-      <tr>
-        <td>Old Password</td>
-        <td><input type=password name=old_password tabindex="1" autofocus></td>
-      </tr>
-      <tr>
-        <td>New Password</td>
-        <td><input type=password name=password tabindex="2"></td>
-      </tr>
-      <tr>
-        <td>New Password (again)</td>
-        <td><input type=password name=password2 tabindex="3"></td>
-      </tr>
-    </table>
-    <input type=submit value=Submit tabindex="4">
-  </form>
-
-{% endblock body %}
